requirements.txt
numpy==1.21.6
pandas==1.4.2
scipy==1.7.2
matplotlib==3.5.0
scikit-learn==1.0.1
statsmodels==0.12.1
quantstats==0.0.59
openpyxl==3.0.10
gcastle==1.0.2
tigramite==5.1.0.0